mcp-server-python/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
bcrypt==4.3.0
box-ai-agents-toolkit==0.0.47
box-sdk-gen==1.17.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
cyclopts==3.22.5
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-users==14.0.1
fastmcp==2.11.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
makefun==1.16.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.7.0
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
parse==1.20.2
passlib==1.7.4
pathable==0.4.4
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.1
